from int to int64

+++ lib/gdo/cluster/gdo_cluster.go
@@ -14,11 +14,11 @@ type Cluster struct {
 }
 
 // Shard :
-func (c *Cluster) Shard(seqID int) (db string) {
+func (c *Cluster) Shard(seqID int64) (db string) {
 	if len(c.Shards) == 0 {
 		panic("gdo> cluster has no shard")
 	}
-	shardNum := len(c.Shards)
+	shardNum := int64(len(c.Shards))
 	for _, shard := range c.Shards {
 		if shard.IsMatch(seqID, shardNum) {
 			db = shard.GetDbName()
@@ -63,13 +63,13 @@ func Init() bool {
 			case "range":
 				shards[i] = &ShardRange{
 					DbName: _cShard["db"].(string),
-					FrNum:  _cShard["fr_num"].(int),
-					ToNum:  _cShard["to_num"].(int),
+					FrNum:  _cShard["fr_num"].(int64),
+					ToNum:  _cShard["to_num"].(int64),
 				}
 			case "hash":
 				shards[i] = &ShardHash{
 					DbName: _cShard["db"].(string),
-					ModRes: _cShard["mod_res"].(int),
+					ModRes: _cShard["mod_res"].(int64),
 				}
 			}
 			i++
